Turn _form_layout into reusable widget
